+++ unlock-js/src/v01/withdrawFromLock.js
@@ -6,8 +6,9 @@ import utils from '../utils'
  * Triggers a transaction to withdraw funds from the lock and assign them to the owner.
  * @param {object} params
  * - {PropTypes.address} lockAddress
+ * @param {function} callback invoked with the transaction hash
  */
-export default async function({ lockAddress }) {
+export default async function({ lockAddress }, callback) {
   const lockContract = await this.getLockContract(lockAddress)
   const transactionPromise = lockContract['withdraw()']({
     gasLimit: GAS_AMOUNTS.withdraw,
@@ -16,6 +17,11 @@ export default async function({ lockAddress }) {
     transactionPromise,
     TransactionTypes.WITHDRAWAL
   )
+
+  if (callback) {
+    callback(null, hash)
+  }
+
   // Let's now wait for the funds to have been withdrawn
   const receipt = await this.provider.waitForTransaction(hash)
   const parser = lockContract.interface